# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Ambil semua produk",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data produk"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tambah produk",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "harga",
                                    "stock"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Keyboard"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 300000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Ambil produk berdasarkan ID",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk ditemukan"
                    },
                    "404": {
                        "description": "Produk tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update semua data produk",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "harga",
                                    "stock"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Monitor"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 2000000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Hapus produk",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk berhasil dihapus"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update stock produk",
                "operationId": "updateProductStock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Produk",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stock"
                                ],
                                "properties": {
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock berhasil diupdate"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "API Products"
        }
    ]
}